requirements.txt
# pip install -r requirements.txt  
# Install dependencies from this file

pandas
python-docx
pdfkit
wkhtmltopdf
fpdf